requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysql-connector-python==9.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
